map editor

+++ source/doodad_brush.h
@@ -83,7 +83,7 @@ protected:
 	bool one_size;
 	bool do_new_borders;
 	bool on_duplicate;
-	uint16_t clear_mapflags;
+	uint32_t clear_mapflags;
 	uint16_t clear_statflags;
 
 	struct SingleBlock {
